clang

+++ fflas-ffpack/fflas/fflas_fgemm.inl
@@ -1171,7 +1171,7 @@ namespace FFLAS {
 				 typename Field::Element* C, const size_t ldc,
 				 const size_t kmax)
 		{
-			size_t mkn = (bool)(nr > 0)+ ((bool)(kr > 0) << 1)+  ((bool)(mr > 0) << 2);
+			size_t mkn =(size_t)( (bool)(nr > 0)+ ((bool)(kr > 0) << 1)+  ((bool)(mr > 0) << 2));
 			if (mkn == 0) return;
 
 			const typename Field::Element *a12, *a21, *b12, *b21;
